a2a-agent/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "agent_server"
version = "0.1.0"
description = "A2A Agent Server for Hotel Booking Management using MCP tools"
requires-python = ">=3.12"
dependencies = [
  "a2a-sdk==0.3.4",
  "mcp>=1.13.1",
  "ollama>=0.5.3",
  "python-dotenv>=1.1.1",
  "httpx>=0.28.1",
  "uvicorn>=0.35.0",
]

[project.scripts]
agent-server = "agent_server.main:main"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent_server"]
